pdf-server/requirements.txt
flask
gunicorn
pypdf
reportlab
Pillow
python-bidi
